+++ package.json
@@ -23,11 +23,13 @@
     "@nestjs/common": "^10.0.0",
     "@nestjs/core": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
+    "@nestjs/schedule": "^4.1.1",
     "@nestjs/swagger": "^7.4.2",
     "dotenv": "^16.4.5",
     "ethers": "^6.13.4",
     "reflect-metadata": "^0.2.0",
-    "rxjs": "^7.8.1"
+    "rxjs": "^7.8.1",
+    "slack-node": "^0.1.8"
   },
   "devDependencies": {
     "@nestjs/cli": "^10.0.0",
