+++ src/modules/block-chain/dtos/block-with-transaction-receipt.dto.ts
