+++ src/providers/ethers/infrastructure/ethers.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { EthersService } from './ethers.service';
+import { EthersService } from '../application/ethers.service';
 
 @Module({
   providers: [EthersService],
-  exports: [EthersService]
+  exports: [EthersService],
 })
 export class EthersModule {}
